requirements.txt
gpt-researcher==0.14.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.5
python-dotenv>=1.1.0
pydantic>=2.9.2
python-multipart>=0.0.20
langchain-mcp-adapters
psutil>=5.9.0